+++ docs/CHANGELOG.rst
@@ -5,6 +5,12 @@ Changelog
 Unreleased
 ----------
 
+* Updated the ignoring warnings in libcli.py
+* Added tests for ignoring warnings
+
+v0.1.34 (2025-11-28)
+------------------------------------------
+
 * Added argcomplete for tab completion
 * Updated docs